refactor(event): add tabs component in ticket management page
