+++ dashboard/src/router.js
@@ -120,6 +120,13 @@ const routes = [
         path: 'tickets',
         name: 'EventTickets',
         component: () => import('@/pages/EventTickets.vue'),
+        children: [
+          {
+            path: '',
+            name: 'EventTicketsManage',
+            component: () => import('@/pages/EventTicketManage.vue'),
+          },
+        ],
       },
       {
         path: 'cfp',
